setRate frame command now works with non-integer values
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@1008 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/anim/anim_blend.cpp
@@ -937,7 +937,7 @@ void idAnim::CallFrameCommands( idEntity *ent, int from, int to, idAnimBlend *ca
 
 				case FC_SETRATE:
 				{
-					int newRate = atof( command.string->c_str() );
+					float newRate = atof( command.string->c_str() );
 					
 					for( int ind = 0; ind < numAnims; ind++ )
 					{
